requirements.txt
nicegui
bcrypt
python-jose[cryptography]
sqlalchemy
bleach
pytest
pytest-cov
mypy
prometheus-client
psutil
slowapi>=0.1.8
itsdangerous>=2.1.0
